requirements.txt
aiofiles
APScheduler
bcrypt
boto3
cryptography
fastapi
google-api-python-client
google_auth_oauthlib
google-cloud-storage
jinja2
oauth2client
oauthlib
opencv-python
passlib
pycryptodome
pyjwt[crypto]
pymongo
pymongo[srv]
python-jose
python_jwt
python-multipart
pytz
razorpay
redis
requests_oauthlib
requests_toolbelt
sentry-sdk
sseclient
uvicorn[standard]